requirements.txt
sep
astropy
pillow
numpy
rawpy
imageio
fitsio
opencv-python